Apply multiple fixes
* Fix `diagnostics` to be empty instead of `null`
* Remove duplicate `BuildSource` call
* Remove `id` validation in `OnInitialized` handler
* Add `--stdio` flag, since VS Code extension adds it
* Update tests

+++ tests/diagnostics-tests.cpp
@@ -77,5 +77,5 @@ TEST(DiagnosticsTest, SelectDeviceBasedOnNonExistingIndex)
     EXPECT_CALL(*mockCLInfo, GetDeviceDescription(_)).WillOnce(Return(""));
 
     auto diagnostics = CreateDiagnostics(mockCLInfo);
-    diagnostics->SetOpenCLDevice(4527288514);
+    diagnostics->SetOpenCLDevice(static_cast<uint32_t>(4527288514));
 }
